+++ ledge_lang/pilot_templates/__init__.py
@@ -1,0 +1,1 @@
+"""Packaged pilot dry-run resources."""